feat: start implementing make transfer
