--- testing/demo/pkg/transfer.go
@@ -1,8 +1,0 @@
-package pkg
-
-import "fmt"
-
-func Transfer() error {
-	fmt.Println("Transferring tokens...")
-	return nil
-}
